Change new() methods to const
This allows for creation of const singleton values, e.g.:
`const INDEX_3: typenum::consts::U3 = typenum::consts::U3::new();`

Due to current restrictions on `const fn` in stable, this requires
the constructed type to not have any trait constraints (other than
Sized), so this commit removes those constraints. This does not break
anything, but may be counter to the initial design / planned
ergonomics of this crate.